Pull Identity By Descent (IBD) haplotypes

Retrieves Identity By Descent (IBD) haplotype data

pullIbdHaplo(
  pop = NULL,
  chr = NULL,
  snpChip = NULL,
  pedigree = NULL,
  simParam = NULL
)

Arguments

pop
an object of Pop-class or RawPop-class. If NULL, haplotypes for the whole ancestral pedigree are retreived. Otherwise, haplotypes just for the pop individuals are retreived. In both cases the base population is controlled by pedigree.

chr
a vector of chromosomes to retrieve. If NULL, all chromosomes are retrieved.

snpChip
an integer. Indicates which SNP array loci are retrieved. If NULL, all sites are retrieved.

pedigree
a matrix with ancestral pedigree to set a base population. It should be of the same form as simParam$pedigree (see setTrackPed in SimParam), i.e., two columns (mother and father) and the same number of rows as simParam$pedigree.Base population can be set by setting parents as 0. If NULL, pedigree from SimParam is taken.

simParam
an object of SimParam

Value

Returns a matrix of haplotypes with Identity By Descent (IBD) coding of locus alleles. The matrix colnames reflect whether all segregagting loci (sites) are retreived or only SNP array loci.

Examples

#Create founder haplotypes
founderPop = quickHaplo(nInd=10, nChr=1, segSites=15)

#Set simulation parameters
SP = SimParam$new(founderPop)
SP$addTraitA(10)
SP$addSnpChip(5)
SP$setTrackRec(TRUE)

#Create population
pop = newPop(founderPop, simParam=SP)
pullIbdHaplo(pop, simParam=SP)

#>      SITE_1 SITE_2 SITE_3 SITE_4 SITE_5 SITE_6 SITE_7 SITE_8 SITE_9 SITE_10
#> 1_1       1      1      1      1      1      1      1      1      1       1
#> 1_2       2      2      2      2      2      2      2      2      2       2
#> 2_1       3      3      3      3      3      3      3      3      3       3
#> 2_2       4      4      4      4      4      4      4      4      4       4
#> 3_1       5      5      5      5      5      5      5      5      5       5
#> 3_2       6      6      6      6      6      6      6      6      6       6
#> 4_1       7      7      7      7      7      7      7      7      7       7
#> 4_2       8      8      8      8      8      8      8      8      8       8
#> 5_1       9      9      9      9      9      9      9      9      9       9
#> 5_2      10     10     10     10     10     10     10     10     10      10
#> 6_1      11     11     11     11     11     11     11     11     11      11
#> 6_2      12     12     12     12     12     12     12     12     12      12
#> 7_1      13     13     13     13     13     13     13     13     13      13
#> 7_2      14     14     14     14     14     14     14     14     14      14
#> 8_1      15     15     15     15     15     15     15     15     15      15
#> 8_2      16     16     16     16     16     16     16     16     16      16
#> 9_1      17     17     17     17     17     17     17     17     17      17
#> 9_2      18     18     18     18     18     18     18     18     18      18
#> 10_1     19     19     19     19     19     19     19     19     19      19
#> 10_2     20     20     20     20     20     20     20     20     20      20
#>      SITE_11 SITE_12 SITE_13 SITE_14 SITE_15
#> 1_1        1       1       1       1       1
#> 1_2        2       2       2       2       2
#> 2_1        3       3       3       3       3
#> 2_2        4       4       4       4       4
#> 3_1        5       5       5       5       5
#> 3_2        6       6       6       6       6
#> 4_1        7       7       7       7       7
#> 4_2        8       8       8       8       8
#> 5_1        9       9       9       9       9
#> 5_2       10      10      10      10      10
#> 6_1       11      11      11      11      11
#> 6_2       12      12      12      12      12
#> 7_1       13      13      13      13      13
#> 7_2       14      14      14      14      14
#> 8_1       15      15      15      15      15
#> 8_2       16      16      16      16      16
#> 9_1       17      17      17      17      17
#> 9_2       18      18      18      18      18
#> 10_1      19      19      19      19      19
#> 10_2      20      20      20      20      20