Select individuals

Selects a subset of nInd individuals from a population.

selectInd(
  pop,
  nInd,
  trait = 1,
  use = "pheno",
  sex = "B",
  selectTop = TRUE,
  returnPop = TRUE,
  candidates = NULL,
  simParam = NULL,
  ...
)

Arguments

pop
and object of Pop-class or HybridPop-class

nInd
the number of individuals to select

trait
the trait for selection. Either a number indicating a single trait or a function returning a vector of length nInd.

use
select on genetic values "gv", estimated breeding values "ebv", breeding values "bv", phenotypes "pheno", or randomly "rand"

sex
which sex to select. Use "B" for both, "F" for females and "M" for males. If the simulation is not using sexes, the argument is ignored.

selectTop
selects highest values if true. Selects lowest values if false.

returnPop
should results be returned as a Pop-class. If FALSE, only the index of selected individuals is returned.

candidates
an optional vector of eligible selection candidates.

simParam
an object of SimParam

...
additional arguments if using a function for trait

Value

Returns an object of Pop-class or HybridPop-class

Examples

#Create founder haplotypes
founderPop = quickHaplo(nInd=10, nChr=1, segSites=10)

#Set simulation parameters
SP = SimParam$new(founderPop)
SP$addTraitA(10)
SP$setVarE(h2=0.5)

#Create population
pop = newPop(founderPop, simParam=SP)

#Select best 5
pop2 = selectInd(pop, 5, simParam=SP)