Population

Extends RawPop-class to add sex, genetic values, phenotypes, and pedigrees.

Usage

# S4 method for Pop
[(x, i)

# S4 method for Pop
c(x, ...)

# S4 method for Pop
show(object)

Arguments

x

a 'Pop' object

i

index of individuals

...

additional 'Pop' objects

object

a 'Pop' object

Methods (by generic)

[: Extract Pop by index or id

c(Pop): Combine multiple Pops

show(Pop): Show population summary

Slots

id

an individual's identifier

iid

an individual's internal identifier

mother

the identifier of the individual's mother

father

the identifier of the individual's father

sex

sex of individuals: "M" for males, "F" for females, and "H" for hermaphrodites

nTraits

number of traits

gv

matrix of genetic values. When using GxE traits, gv reflects gv when p=0.5. Dimensions are nInd by nTraits.

pheno

matrix of phenotypic values. Dimensions are nInd by nTraits.

ebv

matrix of estimated breeding values. Dimensions are nInd rows and a variable number of columns.

gxe

list containing GxE slopes for GxE traits

fixEff

a fixed effect relating to the phenotype. Used by genomic selection models but otherwise ignored.

reps

the number of replications used to measure the phenotype. Used by genomic selection models, but otherwise ignored.

misc

a list whose elements correspond to individuals in the population. This list is normally empty and exists solely as an open slot available for uses to store extra information about individuals.