Create founder haplotypes using MaCS

Uses the MaCS software to produce founder haplotypes.

Usage

runMacs(
  nInd,
  nChr = 1,
  segSites = NULL,
  inbred = FALSE,
  species = "GENERIC",
  split = NULL,
  ploidy = 2L,
  manualCommand = NULL,
  manualGenLen = NULL,
  nThreads = NULL
)

Arguments

nInd

number of individuals to simulate

nChr

number of chromosomes to simulate

segSites

number of segregating sites to keep per chromosome. A value of NULL results in all sites being retained.

inbred

should founder individuals be inbred

species

species history to simulate. See details.

split

an optional historic population split in terms of generations ago.

ploidy

ploidy level of organism

manualCommand

user provided MaCS options. For advanced users only.

manualGenLen

user provided genetic length. This must be supplied if using manualCommand. If not using manualCommand, this value will replace the predefined genetic length for the species. However, this the genetic length is only used by AlphaSimR and is not passed to MaCS, so MaCS still uses the predefined genetic length. For advanced users only.

nThreads

if OpenMP is available, this will allow for simulating chromosomes in parallel. If the value is NULL, the number of threads is automatically detected.

Value

an object of MapPop-class

Details

The current species histories are included: GENERIC, CATTLE, WHEAT, MAIZE, and EUROPEAN.

Examples

# Creates a populations of 10 outbred individuals
# Their genome consists of 1 chromosome and 100 segregating sites
founderPop = runMacs(nInd=10,nChr=1,segSites=100)